# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gorilla Guardians API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List experience packages
      responses:
        "200":
          description: Packages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExperiencePackage"
    post:
      operationId: createPackage
      tags: [packages]
      summary: Create an experience package (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperiencePackageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExperiencePackage"
  /packages/{id}:
    get:
      operationId: getPackage
      tags: [packages]
      summary: Get package by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExperiencePackage"
    patch:
      operationId: updatePackage
      tags: [packages]
      summary: Update package (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperiencePackageInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExperiencePackage"
  /packages/{id}/book:
    post:
      operationId: bookPackage
      tags: [packages]
      summary: Book all experiences included in a package for one date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageBookingInput"
      responses:
        "201":
          description: Bookings created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageBookingResponse"

  # BOOKINGS
...
components:
  schemas:
...
    Experience:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [guided_tour, homestay, workshop, dance_lesson, cooking_class]
        images:
          type: array
          items:
            type: string
        videoUrl:
          type: ["string", "null"]
        price:
          type: number
        duration:
          type: string
        capacity:
          type: integer
        difficultyLevel:
          type: ["string", "null"]
        includedItems:
          type: array
          items:
            type: string
        meetingPoint:
          type: ["string", "null"]
        cancellationPolicy:
          type: ["string", "null"]
        active:
          type: boolean
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, slug, description, type, images, price, duration, capacity, includedItems, active, reviewCount, createdAt]
...
    ExperiencePackage:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
        price:
          type: number
        discountPercent:
          type: number
        active:
          type: boolean
        experiences:
          type: array
          items:
            $ref: "#/components/schemas/Experience"
        createdAt:
          type: string
      required: [id, title, slug, description, images, price, discountPercent, active, experiences, createdAt]
    ExperiencePackageInput:
      type: object
      required: [title, description, price, experienceIds]
      properties:
        title:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
        price:
          type: number
        discountPercent:
          type: number
        active:
          type: boolean
        experienceIds:
          type: array
          items:
            type: integer
          minItems: 2
          description: IDs of the existing experiences bundled into this package, in display order.
    PackageBookingInput:
      type: object
      required: [date, participants]
      properties:
        date:
          type: string
          format: date
        participants:
          type: integer
          minimum: 1
        specialRequests:
          type: string
    PackageBookingResponse:
      type: object
      properties:
        packageBookingRef:
          type: string
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
      required: [packageBookingRef, bookings]
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        experienceId:
          type: integer
        date:
          type: string
          format: date
        participants:
          type: integer
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        paymentStatus:
          type: string
          enum: [pending, paid, refunded]
        specialRequests:
          type: ["string", "null"]
        experience:
          $ref: "#/components/schemas/Experience"
        createdAt:
          type: string
          format: date-time
      required: [id, userId, experienceId, date, participants, totalAmount, status, paymentStatus, createdAt]
...