# lib/api-zod/src/generated/types/experiencePackageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gorilla Guardians API
 * OpenAPI spec version: 0.1.0
 */

export interface ExperiencePackageInput {
  title: string;
  description: string;
  images?: string[];
  price: number;
  discountPercent?: number;
  active?: boolean;
  /**
     * IDs of the existing experiences bundled into this package, in display order.
     * @minItems 2
     */
  experienceIds: number[];
}