# lib/api-zod/src/generated/types/packageBookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gorilla Guardians API
 * OpenAPI spec version: 0.1.0
 */

export interface PackageBookingInput {
  date: Date;
  /** @minimum 1 */
  participants: number;
  specialRequests?: string;
}