# lib/api-zod/src/generated/types/packageBookingResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gorilla Guardians API
 * OpenAPI spec version: 0.1.0
 */
import type { Booking } from './booking';

export interface PackageBookingResponse {
  packageBookingRef: string;
  bookings: Booking[];
}